Spilt OAuth classes into separate files
